Converted to 3 separate view models. Wired with statistics screen.
